정수처리 챗봇/requirements.txt
# 웹 서버 및 API
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.2
python-multipart>=0.0.6

# 기본 유틸리티
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.3,<2.0
pandas>=2.1.4

# PDF 처리
PyPDF2>=3.0.1
pymupdf>=1.23.14
pdfplumber>=0.10.4

# OCR 및 이미지 처리 (PDF 텍스트 인식률 향상)
pytesseract>=0.3.10
Pillow>=10.1.0
opencv-python>=4.8.1.78,<4.10

# 벡터 저장소 및 임베딩
sentence-transformers>=2.2.2,<2.3
faiss-cpu>=1.7.4
chromadb>=0.4.18
huggingface-hub>=0.4.0,<0.20

# AI 모델 및 머신러닝 (GPU 지원)
transformers>=4.36.2,<4.40
tokenizers>=0.15.0
torch>=2.1.0,<2.2
torchvision>=0.16.0,<0.17

# GPU 지원 추가 패키지
accelerate>=0.25.0
bitsandbytes>=0.41.1,<0.42

# 자연어 처리 및 텍스트 분석
spacy>=3.7.0
scikit-learn>=1.3.0,<1.4

# 데이터 처리 및 분석
scipy>=1.11.0

# 시스템 모니터링
psutil>=5.9.6

# 로깅
loguru>=0.7.2

# 추가 유틸리티
pathlib2>=2.3.7
typing-extensions>=4.8.0
dataclasses-json>=0.6.3

# 개발 및 테스트 도구
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

